Catch filemanager-gksudo2 set as default for unknown files
With some browsers (pcmanfm, for example) an unreadable file me be reported as "unknown" type, and xdg mime will not show a default to open unknown unreadable files. Yet these are treated as "application/octet-stream" by gio and elsewhere and filemanager-gksudo2 may be default for this type. Correct script to deal with this.